requirements.txt
requests
pandas
beautifulsoup4

requests==2.32.3
feedparser==6.0.11
python-dateutil==2.9.0.post0